split maintoolbar into pages as well
