+++ src/client/app/reducers/rootReducer.jsx
@@ -3,6 +3,7 @@ import { combineReducers } from 'redux';
 import editorContainer from './editorContainer.jsx';
 import iframe from './iframe.jsx';
 import mainToolbar from './mainToolbar.jsx';
+import page from './page.jsx';
 import textEditors from './textEditors.jsx';
 import user from './user.jsx';
 
@@ -10,6 +11,7 @@ const rootReducer = combineReducers({
   editorContainer,
   iframe,
   mainToolbar,
+  page,
   textEditors,
   user
 });
